Fix check on synchronized modules
Base setting if I_MU_PHSP (scoring of MU index
in IAEA format phase space files) on whether
characters 1-4 of CMNAME for any module in the
accelerator are 'SYNC'.  Replaces check on full
value of CMNAME.  The latter check failed if the only
synchronized CM was SYNCHDMLC, in which case
CMNAME can never be equal to 'SYNCHDMLC' because
CMNAME is character*8!

Resolves Issue #533